create a tennis player with repeat
This was one of the first C++ programs I made back in 2017 when I was starting to learn C++. One of the problems with this code is that it can take a one word name, but type a first and last name and the program crashes. However, this does show how to ask the user for basic string and int input.